requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-paypal==0.4.1
idna==2.6
Pillow==5.2.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
